Revert "Preserve source strings when finding mismatches."
This reverts commit 146b9fed3bdb11e2eb62ff4e5eb9731a116c1b5a.

This way of doing things does not work out super cleanly because CSV files can have separate inflow and outflow columns that get summed together, so we'd need to "sum together" the source strings, and that seems bad. We'd better just have a dropdown in the UI somewhere to choose the display format.