WIP: Experiment with controlling the file input.
This lets us add a placeholder filename when the form is reopened (we show the filename of whatever is currently imported).

This works, but it raises a UX question of what we want to happen when the import form is reopened. Do we want the user to start from scratch so we can progressively show them the instructions again, or do we want them to be able to